% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_ci_median.R
\name{div_ci_median}
\alias{div_ci_median}
\title{Function to calculate the confidence interval for the median}
\usage{
div_ci_median(x, conf = 0.95)
}
\arguments{
\item{x}{numeric, data from which the median is calcualted}

\item{conf}{numeric, the confidence interval as 1 - P(x < x0)}
}
\value{
ci (confidence interval object)
}
\description{
Function to calculate the confidence interval for the median
}
\examples{
x <- 1:100
div_ci_median(x)
}
