% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_conf_colour.R
\name{div_conf_colour}
\alias{div_conf_colour}
\title{return a colour code given a number of stars for the confidence level of bias}
\usage{
div_conf_colour(x)
}
\arguments{
\item{x}{the string associated to the paygap confidence: NA, '', ',', '*', '***', '***'}
}
\value{
string (named colour)
}
\description{
This function returns a colour (R named colour) based on the confidence level
}
\examples{
div_conf_colour("*")

}
\keyword{parse}
\keyword{paygap}
