% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_plot_paygap_distribution.R
\name{div_plot_paygap_distribution}
\alias{div_plot_paygap_distribution}
\title{Produce a histogram and normal distribution}
\usage{
div_plot_paygap_distribution(x, label = "Gender", mu_unbiased = 1)
}
\arguments{
\item{x}{numeric vector, column of paygap observations}

\item{label}{character, prefix for the title}

\item{mu_unbiased}{numeric, the mean of the unbiased distribution (for paygaps this should be 1)}
}
\value{
ggplot2 object
}
\description{
Plots a histogram, a normal distribution with the same standard deviation and mean as well as one with a mean centred around 1
}
\examples{
d <- div_fake_team()
pg <- div_paygap(d)
div_plot_paygap_distribution(pg$data$paygap)

}
\keyword{distribution}
\keyword{equity}
