% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_round_paygap.R
\name{div_round_paygap}
\alias{div_round_paygap}
\title{Rounds all numbers in the paygap data-frame}
\usage{
div_round_paygap(x)
}
\arguments{
\item{x}{paygap object (output of div::div_paygap())}
}
\value{
the paygap data-frame (tibble only, not the whole paygap object)
}
\description{
This function all numbers to zero decimals, except the paygap (which is rounded to 2 decimals):
}
\examples{
d <- div_fake_team()
pg <- div_paygap(d)
div_round_paygap(pg)
}
