% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Calculate the diversity index}
\usage{
diversity(x, prior = NULL)
}
\arguments{
\item{x}{numeric vector, observed probabilities of the classes}

\item{prior}{numeric vector, the prior probabilities of the classes}
}
\value{
the entropy or diversity measure
}
\description{
This function calculates the entropy of a system with discrete states
}
\examples{
x <- c(0.4, 0.6)
diversity(x)
}
\keyword{entropy}
