% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.paygap.R
\name{print.paygap}
\alias{print.paygap}
\title{print the paygap object in the terminal}
\usage{
\method{print}{paygap}(x, ...)
}
\arguments{
\item{x}{paygap object, as created by the function div_paygpa()}

\item{...}{arguments passed on to the generic print function: print(x$data)}
}
\value{
text output
}
\description{
print the paygap object in the terminal
}
\examples{
library(div)
div_fake_team() \%>\%
  div_paygap    \%>\%
  print
}
